+++ src/dipdup/cli.py
@@ -340,19 +340,22 @@ async def config(ctx) -> None:
 
 @config.command(name='export')
 @click.option('--unsafe', is_flag=True, help='Resolve environment variables or use default values from config.')
+@click.option('--full', is_flag=True, help='Resolve index templates.')
 @click.pass_context
 @cli_wrapper
-async def config_export(ctx, unsafe: bool) -> None:
+async def config_export(ctx, unsafe: bool, full: bool) -> None:
     """
-    Print config after resolving all links and templates.
+    Print config after resolving all links and, optionally, templates.
 
     WARNING: Avoid sharing output with 3rd-parties when `--unsafe` flag set - it may contain secrets!
     """
-    config_yaml = DipDupConfig.load(
+    config = DipDupConfig.load(
         paths=ctx.obj.config.paths,
         environment=unsafe,
-    ).dump()
-    echo(config_yaml)
+    )
+    if full:
+        config.initialize(skip_imports=True)
+    echo(config.dump())
 
 
 @config.command(name='env')